requirements.txt
pwntools
paramiko
sys
optparse